move: move hooks folder into src from components folder
